refactor: shift actions component
